Changed values for user balance and fees to DECIMAL type

+++ models.py
@@ -1,5 +1,5 @@
 from config import Config
-from sqlalchemy import (Column, Integer, String, Boolean, BigInteger,
+from sqlalchemy import (Column, Integer, String, Boolean, BigInteger, DECIMAL,
                         Float, Date, DateTime, ForeignKey, create_engine)
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.sql import func
@@ -23,7 +23,7 @@ class User(Base, UserMixin):
     userpass = Column(String(200), unique=False, nullable=False)
     verified = Column(Boolean, unique=False, nullable=False)
     banned = Column(Boolean, unique=False, nullable=False)
-    balance = Column(Float, unique=False, nullable=False)
+    balance = Column(DECIMAL(20, 2), unique=False, nullable=False)
 
     def get_id(self):
         return self.userID
@@ -39,8 +39,8 @@ class Usershare(Base):
                     primary_key=True)
     issuerID = Column(String(3), ForeignKey('SHARE.issuerID'),
                       primary_key=True)
-    profit = Column(Float, nullable=False, unique=False)
-    loss = Column(Float, nullable=False, unique=False)
+    profit = Column(DECIMAL(20, 2), nullable=False, unique=False)
+    loss = Column(DECIMAL(20, 2), nullable=False, unique=False)
     quantity = Column(BigInteger, nullable=False, unique=False)
 
 
@@ -55,9 +55,9 @@ class Transaction(Base):
     userID = Column(Integer, ForeignKey('USER.userID'), primary_key=False)
     datetime = Column(DateTime, nullable=False, unique=False)
     transtype = Column(String(1), nullable=False, unique=False)
-    feeval = Column(Float, nullable=False, unique=False)
-    stocktransval = Column(Float, nullable=False, unique=False)
-    totaltransval = Column(Float, nullable=False, unique=False)
+    feeval = Column(DECIMAL(20, 2), nullable=False, unique=False)
+    stocktransval = Column(DECIMAL(20, 2), nullable=False, unique=False)
+    totaltransval = Column(DECIMAL(20, 2), nullable=False, unique=False)
     quantity = Column(BigInteger, nullable=False, unique=False)
     status = Column(String(20), nullable=False)
